+++ resources/lang/sk/informacie.php
@@ -26,13 +26,17 @@ return array(
         'SGB' => 'Stredoslovenská galéria',
         'GUS' => 'Galéria umelcov Spiša',
         'VSG' => 'Východoslovenská galéria',
-        'MG' => 'Moravská galerie',
         'TGP' => 'Tatranská galéria',
         'PGU' => 'Považská galéria umenia',
         'SGP' => 'Šariššská galéria v Prešove',
         'KGC' => 'Kysucká galéria v Oščadnici',
+        'TGM' => 'Turčianska galéria',
+        'GJK' => 'Galéria Jána Koniarka',
+        'MVL' => 'Múzeum Vojtecha Löfflera',
+        'MG' => 'Moravská galerie',
         'PNP' => 'Památník národního písemnictví',
         '4RG' => 'Galerie moderního umění v Roudnici nad Labem',
+        'GMUHK' => 'Galerie moderního umění v Hradci Králové',
     ],
 
     'mozete_heading'               => 'U nás môžete',
@@ -56,16 +60,16 @@ return array(
     'more-items_connect_col2_lead'       => 'sme otvorení <strong>ďalším <br>formám&nbsp;spolupráce</strong>.',
     'more-items_connect_col2_p'          => 'Ak viete o niektorom z diel alebo autorov viac informácií, alebo ste na jeho profile u nás nenašli fotografiu, budeme radi ak nám pomôžete pri ich doplnení. Napíšte nám na <a href="mailto:lab@sng.sk">lab@sng.sk</a>, spojíme sa s vami.',
     'more-items_connect_col3_lead'       => 'web umenia realizuje <br>kolektív <strong><a href="http://lab.sng.sk/" target="_blank" class="strong">lab.SNG</a></strong>',
-    'more-items_connect_col3_ul-content' => '<li><strong>Michal Čudrnák</strong> vedúci lab.SNG, autor článkov</li>
-                                             <li><strong>Lukáš Štepanovský</strong> editor obsahu, autor článkov, UX dizajnér</li>
+    'more-items_connect_col3_ul-content' => '<li><strong>Michal Čudrnák</strong> vedúci lab.SNG</li>
+                                             <li><strong>Lukáš Štepanovský</strong> editor obsahu, UX dizajnér</li>
                                              <li><strong>Igor Rjabinin</strong> vývojár</li>
                                              <li><strong>Rastislav Chynoranský</strong> vývojár</li>
                                              <li><strong>Ernest Walzel</strong> vývojár</li>
-                                             <li><strong>Martin Havala</strong> vývojár</li>
-                                             <li><strong>Andrej Bilec</strong> vývojár</li>
-                                             <li><strong>Zuzana Koblišková</strong> editorka obsahu, autorka článkov</li>
-                                             <li><strong>Gabriel Balkó</strong> kontrolór kvality, editor obsahu</li>
-                                             <li><strong>Karin Vicianová</strong> kontrolórka kvality, editorka obsahu</li>
+                                             <li><strong>Gorazd Ratulovský</strong> UX dizajnér</li>
+                                             <li><strong>Katarína Buzová</strong> UX výskumníčka</li>
+                                             <li><strong>Zuzana Koblišková</strong> editorka obsahu</li>
+                                             <li><strong>Gabriel Balkó</strong> kontrolór kvality, správca reprodukcií</li>
+                                             <li><strong>Karin Vicianová</strong> kontrolórka kvality</li>
                                              <li><strong>Jana Šuchová</strong> editorka obsahu</li>
                                              <li><strong>Vladimír Hergott</strong> správca CEDVU</li>
                                              <li><strong>Braňo Matis</strong> grafický dizajnér</li>',