+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 All notable changes to this project will be documented in this file[^1].
 
 ## [Unreleased]
-
+### Changed
+- Removed PNP from downloadable items
+- Information page with updated data
+-
 ## [2.53.0] - 2022-05-13
 ### Changed
 - temporarily removed print reproductions